Add Edit Modal & other fixes

+++ app.js
@@ -1,4 +1,5 @@
 let addBtn = document.getElementById('add');
+let updateBtn = document.getElementById('update');
 let clearBtn = document.getElementById('clear');
 let data = {};
 let userObj = [];
@@ -12,7 +13,7 @@ const maskPassword = (password) => {
 }
 
 const resetFormData = () => {
-    document.getElementById('userForm').reset();
+    document.getElementById('userAddForm').reset();
 }
 
 const getData = () => {
@@ -37,8 +38,9 @@ const getData = () => {
                     <td>${index+1}</td>
                     <td>${obj.username}</td>
                     <td>${maskPassword(obj.password)}</td>
-                    <td><a href="${obj.website}">${obj.website}</a></td>
+                    <td><a href="${obj.website}" target="_blank">${obj.website}</a></td>
                     <td>
+                        <button class="action--btn edit" data-ind="${index}">Edit</button>
                         <button class="action--btn copy" data-ind="${index}">Copy</button>
                         <button class="action--btn delete" data-ind="${index}">Delete</button>
                     </td>
@@ -56,12 +58,23 @@ const getData = () => {
                     }, 2000);
                 })
             })
+
+            Array.from(document.querySelectorAll('button.edit')).forEach(element => {
+                element.addEventListener('click', (e) => {
+                    const ind = e.target.dataset.ind;
+                    const user = userObj[ind];
+
+                    toggleEditModal(ind, user.password, user.website);
+                })
+            })
             
             Array.from(document.querySelectorAll('button.delete')).forEach(element => {
                 element.addEventListener('click', (e) => {
                     const ind = e.target.dataset.ind;
-                    deletePassword(ind);
-                    getData();
+                    if (window.confirm(`Do you want to delete ${userObj[ind].username}'s data ?`)) {
+                        deletePassword(ind);
+                        getData();
+                    }
                 })
             })
         }
@@ -82,7 +95,7 @@ const addPassword = () => {
             return e.username === username.value
         });
     
-        if (!d) {
+        if (!d || d.website != website.value) {
             userObj.push({
                 username: username.value,
                 password: password.value,
@@ -92,18 +105,23 @@ const addPassword = () => {
             getData();
             resetFormData();
         } else {
-            // alert("This data is already present!")
-            if (window.confirm("This user's data already exists! Do you want to update the record ?")) {
-                const ind = userObj.findIndex(e => {
-                    return e.username === username.value;
-                })
-                userObj[ind].password = password.value;
-                userObj[ind].website = website.value;
-                localStorage.setItem('passX', JSON.stringify(userObj));
-                getData();
-                resetFormData();
-            }
+            alert("This data is already exists!")
         }
+    }
+}
+
+const updatePassword = () => {
+    let userID = document.getElementById('userID');
+    let password = document.getElementById('passwordEdit');
+    let website = document.getElementById('websiteEdit');
+
+    if (password.value === "" || website.value === "") {
+        alert("Please fill up your details!")
+    } else {
+        userObj[userID.value].password = password.value;
+        userObj[userID.value].website = website.value;
+
+        localStorage.setItem('passX', JSON.stringify(userObj));
     }
 }
 
@@ -121,7 +139,41 @@ addBtn.addEventListener('click', () => {
     addPassword();
 })
 
-clearBtn.addEventListener('click', () => {
-    localStorage.removeItem('passX');
+updateBtn.addEventListener('click', () => {
+    updatePassword();
+    let userIdEdit = document.getElementById('userID');
+    document.getElementById('userEditForm').reset();
+    userIdEdit.value = "";
+
+    editModal.style.display = 'none';
     getData();
+})
+
+clearBtn.addEventListener('click', () => {
+    if (window.confirm('Do you want to clear all users data ?')) {
+        localStorage.removeItem('passX');
+        getData();
+    }
+})
+
+// Modal handlers
+let editModal = document.getElementById('modal');
+let closeBtn = document.getElementById('closeBtn');
+
+const toggleEditModal = (userID, password, website) => {
+    let userIdEdit = document.getElementById('userID');
+    let websiteEdit = document.getElementById('websiteEdit');
+
+    userIdEdit.value = userID;
+    passwordEdit.value = password;
+    websiteEdit.value = website;
+    editModal.style.display = 'block';
+}
+
+closeBtn.addEventListener('click', () => {
+    let userIdEdit = document.getElementById('userID');
+    document.getElementById('userEditForm').reset();
+    userIdEdit.value = "";
+
+    editModal.style.display = 'none';
 })
